consolidate themes into one file
