+++ frontend/src/components/GenericModal.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import Modal from "@mui/material/Modal";
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
-import {lightTheme} from "../style/lightTheme";
+import {lightTheme} from "../style/themes";
 
 const modalStyle = {
 	position: "absolute",
