+++ frontend/src/components/mainLandingVisual.tsx
@@ -3,7 +3,7 @@ import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import {ThemeProvider} from "@mui/material/styles";
-import {lightTheme} from "../style/lightTheme";
+import {lightTheme} from "../style/themes";
 
 import {CONFIG} from "../config";
